+++ test/tool/integrationtest/test-issues/testdynamicparameter/source/class/testdynamicparameter/theme/Theme.js
@@ -1,0 +1,21 @@
+/* ************************************************************************
+
+   Copyright: 2025 
+
+   License: MIT license
+
+   Authors: 
+
+************************************************************************ */
+
+qx.Theme.define("testdynamicparameter.theme.Theme",
+{
+  meta :
+  {
+    color : testdynamicparameter.theme.Color,
+    decoration : testdynamicparameter.theme.Decoration,
+    font : testdynamicparameter.theme.Font,
+    icon : qx.theme.icon.Tango,
+    appearance : testdynamicparameter.theme.Appearance
+  }
+});