+++ test/tool/integrationtest/test-translation.js
@@ -16,16 +16,22 @@ async function simpleParsePo(filename) {
   return po;
 }
 
-test("test translation file update", async assert => {
-  try {
-    await testUtils.deleteRecursive("test-translation/tranapp/compiled");
-    await testUtils.safeDelete("test-translation/tranapp/source/translation/en.po");
-    await fsPromises.writeFile("test-translation/tranapp/source/translation/en.po",
+async function prepare() {
+   await testUtils.deleteRecursive("test-translation/tranapp/compiled");
+   await testUtils.safeDelete("test-translation/tranapp/source/translation/en.po");
+   await fsPromises.writeFile("test-translation/tranapp/source/translation/en.po",
 `
 msgid "Lib Override"
 msgstr "lib-override-replaced-value"
 `, "utf8");
-    await testUtils.runCompiler("test-translation/tranapp", "-u");
+}
+
+test("test translation file update", async assert => {
+  try {
+    let result;
+    await prepare();
+    result = await testUtils.runCompiler("test-translation/tranapp", "-u");
+    assert.ok(result.exitCode === 0, testUtils.reportError(result));
     let po = await simpleParsePo("test-translation/tranapp/source/translation/en.po");
     assert.ok(po["App Alpha"] !== undefined);
     assert.ok(po["App Beta"] !== undefined);
@@ -34,18 +40,21 @@ msgstr "lib-override-replaced-value"
     assert.ok(po["Lib Beta"] === undefined);
     assert.ok(po["Lib Charlie"] === undefined);
 
-    await testUtils.runCompiler("test-translation/tranapp", "-u", "--library-po=untranslated");
+    await prepare();
+    result = await testUtils.runCompiler("test-translation/tranapp", "-u", "--library-po=untranslated");
+    assert.ok(result.exitCode === 0, testUtils.reportError(result));
     po = await simpleParsePo("test-translation/tranapp/source/translation/en.po");
-    assert.ok(po["Lib Alpha"] === undefined);
-    assert.ok(po["Lib Beta"] !== undefined);
-    assert.ok(po["Lib Charlie"] !== undefined);
+    assert.ok(po["Lib Alpha"] === undefined, po["Lib Alpha"]);
+    assert.ok(po["Lib Beta"] !== undefined, po["Lib Beta"]);
+    assert.ok(po["Lib Charlie"] !== undefined, po["Lib Charlie"]);
 
-    await testUtils.runCompiler("test-translation/tranapp", "-u", "--library-po=all");
+    await prepare();
+    result = await testUtils.runCompiler("test-translation/tranapp", "-u", "--library-po=all");
+    assert.ok(result.exitCode === 0, testUtils.reportError(result));
     po = await simpleParsePo("test-translation/tranapp/source/translation/en.po");
-    assert.ok(po["Lib Alpha"] === "lib-alpha-value");
-    assert.ok(po["Lib Beta"] !== undefined);
-    assert.ok(po["Lib Charlie"] !== undefined);
-    assert.ok(po["Lib Charlie"] !== undefined);
+    assert.ok(po["Lib Alpha"] === "lib-alpha-value", po["Lib Alpha"]);
+    assert.ok(po["Lib Beta"] !== undefined, po["Lib Beta"]);
+    assert.ok(po["Lib Charlie"] !== undefined, po["Lib Charlie"]);
 
     let data = await fsPromises.readFile("test-translation/tranapp/compiled/source/tranapp/package-0.js", "utf8");
     assert.ok(!!data.match(/lib-override-replaced-value/));