+++ test/tool/integrationtest/testlib/compile.js
@@ -2,7 +2,7 @@ const fs = require("fs");
 const path = require("path");
 
 qx.Class.define("qxl.compilertests.testlib.CompilerApi", {
-  extend: qx.tool.cli.api.CompilerApi,
+  extend: qx.tool.compiler.cli.api.CompilerApi,
   
   members: {
     async load() {
@@ -16,7 +16,7 @@ qx.Class.define("qxl.compilertests.testlib.CompilerApi", {
 });
 
 qx.Class.define("qxl.compilertests.testlib.LibraryApi", {
-  extend: qx.tool.cli.api.LibraryApi,
+  extend: qx.tool.compiler.cli.api.LibraryApi,
   
   members: {
     async load() {